code/requirements.txt
influxdb
requests
opentsdb-py
psycopg2-binary
httplib2
timeout-decorator
boto3
celery
plotly
pandas